Moved business rules output processing a level up.
Processing was bound to Service class, but hosts may also have business
rules defined as check_command. So all business output processing
methods have been moved into the parent SchedulingItem class.
