+++ shinken/objects/schedulingitem.py
@@ -30,6 +30,7 @@ will find all scheduling related functions, like the schedule
 or the consume_check. It's a very important class!
 """
 
+import re
 import random
 import time
 import traceback
@@ -1387,6 +1388,100 @@ class SchedulingItem(Item):
             #print "got node", node
             self.business_rule = node
 
+
+    # Returns a status string for business rules based services, formatted
+    # using format string as output template.
+    # The template may embed output formatting for itself, and for its child
+    # (dependant) itmes. Childs format string is expanded into the $( and )$,
+    # using the string between brackets as format string.
+    # Output is generated by expanding the following macros.
+    #
+    #   SERVICE_DESCRIPTION:    The service name
+    #   HOST_NAME:              The item name
+    #   FULL_NAME:              The item name
+    #   STATUS:                 The item status string (may be OK, WARN, CRIT,
+    #                           UP, DOWN)
+    #   SHORT_STATUS:           The item shorten status string (O, W, C, U, D)
+    #
+    # Caution: only childs in state not OK are displayed.
+    #
+    # Example:
+    #   A business rule with a format string looking like
+    #       "$STATUS$ [ $($STATUS$: $HOST_NAME$,$SERVICE_DESCRIPTION$ )$ ]"
+    #   Would return
+    #       "CRITICAL [ CRITICAL: host1,srv1 WARNING: host2,srv2  ]"
+    def get_business_rule_output(self):
+        got_business_rule = hasattr(self, 'got_business_rule') and \
+                            self.got_business_rule is True
+        # Checks that the service is a business rule.
+        if got_business_rule is False or self.business_rule is None:
+            return ""
+        # Checks that the business rule has a format specified.
+        output_template = self.business_rule_output_template
+        if not output_template:
+            return ""
+        # No child output if business rule state is OK
+        if self.business_rule.get_state() == 0:
+            return "OK: all checks were successful."
+
+        # Extracts template strings
+        # Current service output format string
+        service_template_string = re.sub("\$\(.*\)\$", "$CHILDS_OUTPUT$", output_template)
+        # Childs output format string
+        elts = re.findall(r"\$\((.*)\)\$", output_template)
+        if not len(elts):
+            child_template_string = ""
+        else:
+            child_template_string = elts[0]
+
+        # Processes child services output
+        childs_output = ""
+        # Expands child items format string macros.
+        for child in self.business_rule.list_all_elements():
+            # Do not display childs in OK state
+            if child.last_hard_state_id == 0:
+                continue
+            childs_output += self.expand_business_rule_item_macros(child_template_string, child)
+
+        # Expands node's template string macros.
+        # State has to be set manually, as the service state attribute is only
+        # set on a next scheduler step.
+        output = service_template_string
+        mapping = {0: "OK", 1: "WARNING", 2: "CRITICAL"}
+        status = mapping[self.business_rule.get_state()]
+        output = re.sub(r"\$STATUS\$", status, output, flags=re.I)
+        short_status = self.status_to_short_status(status)
+        output = re.sub(r"\$SHORT_STATUS\$", short_status, output, flags=re.I)
+        output = self.expand_business_rule_item_macros(output, self)
+        output = re.sub("\$CHILDS_OUTPUT\$", childs_output, output)
+        return output
+
+    # Expands format string macros with item attributes
+    def expand_business_rule_item_macros(self, template_string, item):
+        output = template_string
+        status = getattr(item, "state", "")
+        output = re.sub(r"\$STATUS\$", status, output, flags=re.I)
+        short_status = self.status_to_short_status(status)
+        output = re.sub(r"\$SHORT_STATUS\$", short_status, output, flags=re.I)
+        service_description = getattr(item, "service_description", "")
+        output = re.sub(r"\$SERVICE_DESCRIPTION\$", service_description, output, flags=re.I)
+        host_name = getattr(item, "host_name", "")
+        output = re.sub(r"\$HOST_NAME\$", host_name, output, flags=re.I)
+        full_name = item.get_full_name()
+        output = re.sub(r"\$FULL_NAME\$", full_name, output, flags=re.I)
+        return output
+
+    # Returns status string shorten name.
+    def status_to_short_status(self, status):
+        mapping = {
+            "OK": "O",
+            "WARNING": "W",
+            "CRITICAL": "C",
+            "UNKNOWN": "U",
+            "UP": "U",
+            "DOWN": "D"
+        }
+        return mapping.get(status, status)
 
     # We ask us to manage our own internal check,
     # like a business based one
